+++ cli/src/pkg/config_validate.py
@@ -2,140 +2,44 @@
 
 Used by 'dtaas admin config validate'. Each check returns a list of human
 readable problems (empty when acceptable); validate_config aggregates them so
-the user sees every issue at once rather than one at a time.
+the user sees every issue at once rather than one at a time. See
+validators.py for the generic, DTaaS-agnostic predicates each check builds on.
 """
 
-import ipaddress
-import re
-from pathlib import Path, PurePosixPath, PureWindowsPath
-from email_validator import EmailNotValidError, validate_email
-from fqdn import FQDN
 from . import utils
-from .constants import USERNAME_RE
-
-
-URL_RE = re.compile(r"^https?://[A-Za-z0-9.-]+(:\d+)?(/[A-Za-z0-9._~%/+-]*)?$")
-SIZE_RE = re.compile(r"^\d+(\.\d+)?\s*([kmgt]i?b?|b)$", re.IGNORECASE)
-NUMERIC_HOST_RE = re.compile(r"^[0-9.]+$")
+from .validators import (
+    get_nested,
+    is_email,
+    is_existing_dir,
+    is_host,
+    is_int,
+    is_number,
+    is_size,
+    is_string_list,
+    is_url,
+    is_username,
+    optional,
+    required,
+)
 
 _SIZE_FIELDS = (("mem_limit", "4G"), ("shm_size", "512m"))
 
 
-def _get(data, *keys):
-    """Return the nested value at *keys*, or None if any step is missing."""
-    node = data
-    for key in keys:
-        node = node.get(key) if isinstance(node, dict) else None
-    return node
-
-
-def _is_url(value):
-    """True when *value* is an http(s) URL (case-insensitive) with a host/path."""
-    return isinstance(value, str) and bool(URL_RE.match(value.lower()))
-
-
-def _is_ip(value):
-    """True when *value* is a valid IPv4 or IPv6 address."""
-    try:
-        ipaddress.ip_address(value)
-        return True
-    except ValueError:
-        return False
-
-
-def _is_fqdn(value):
-    """True when *value* is an RFC 1123 fully qualified domain name."""
-    return FQDN(value).is_valid
-
-
-def _is_host(value):
-    """True for 'localhost', an IP literal, or a dotted FQDN (numeric => IP)."""
-    if not isinstance(value, str):
-        return False
-    if NUMERIC_HOST_RE.match(value):
-        return _is_ip(value)
-    return value == "localhost" or _is_fqdn(value)
-
-
-def _is_abs_path(value):
-    """True when *value* is an absolute path in POSIX or Windows form."""
-    if not isinstance(value, str) or not value.strip():
-        return False
-    return PurePosixPath(value).is_absolute() or PureWindowsPath(value).is_absolute()
-
-
-def _is_existing_dir(value):
-    """True when *value* is an absolute path to a directory that exists."""
-    return _is_abs_path(value) and Path(value).is_dir()
-
-
-def _is_int(value):
-    """True when *value* is an integer (and not a bool, which subclasses int)."""
-    return isinstance(value, int) and not isinstance(value, bool)
-
-
-def _is_number(value):
-    """True when *value* is a positive number of cores (int or float, not bool)."""
-    if isinstance(value, bool) or not isinstance(value, (int, float)):
-        return False
-    return value > 0
-
-
-def _is_size(value):
-    """True when *value* is a byte size with a required unit, e.g. '4G' or '512m'."""
-    return isinstance(value, str) and bool(SIZE_RE.match(value))
-
-
-def _is_email(value):
-    """True when *value* is a valid email address (no DNS lookup)."""
-    try:
-        validate_email(value, check_deliverability=False)
-        return True
-    except (EmailNotValidError, TypeError):  # TypeError: non-string value
-        return False
-
-
-def _is_username(value):
-    """True when *value* is a non-empty username (alphanumeric plus . _ -)."""
-    return isinstance(value, str) and bool(USERNAME_RE.match(value))
-
-
-def _required(data, keys, rule):
-    """Return [message] if the value at *keys* is missing or fails *rule*.
-
-    *rule* is a (predicate, message) pair.
-    """
-    predicate, message = rule
-    value = _get(data, *keys)
-    if value is None:
-        return [".".join(keys) + " is missing"]
-    return [] if predicate(value) else [message]
-
-
-def _optional(data, keys, rule):
-    """Like _required, but a missing value is allowed (no error)."""
-    predicate, message = rule
-    value = _get(data, *keys)
-    if value is None:
-        return []
-    return [] if predicate(value) else [message]
-
-
 def _check_git_repo(data):
     """git-repo must be a URL."""
-    return _required(data, ("git-repo",), (_is_url, "git-repo must be a valid URL"))
+    return required(data, ("git-repo",), (is_url, "git-repo must be a valid URL"))
 
 
 def _check_server_dns(data):
     """common.server-dns must be a hostname or IP."""
     message = "common.server-dns must be a valid hostname or IP address"
-    return _required(data, ("common", "server-dns"), (_is_host, message))
+    return required(data, ("common", "server-dns"), (is_host, message))
 
 
 def _check_path(data):
     """common.path must be an absolute path to an existing directory."""
     message = "common.path must be an absolute path to an existing directory"
-    return _required(data, ("common", "path"), (_is_existing_dir, message))
+    return required(data, ("common", "path"), (is_existing_dir, message))
 
 
 def _check_certs_src(data):
@@ -144,7 +48,7 @@ def _check_certs_src(data):
     message = (
         "common.security.certs-src must be an absolute path to an existing directory"
     )
-    return _optional(data, keys, (_is_existing_dir, message))
+    return optional(data, keys, (is_existing_dir, message))
 
 
 def _check_resources(data):
@@ -153,35 +57,30 @@ def _check_resources(data):
     With set_limits = false the container runs uncapped, so the fields are
     optional; any values that are present are still checked.
     """
-    enabled = _get(data, "common", "resources", "set_limits") is not False
-    check = _required if enabled else _optional
+    enabled = get_nested(data, "common", "resources", "set_limits") is not False
+    check = required if enabled else optional
     cpus_msg = "common.resources.cpus must be a positive number of CPU cores"
-    errors = check(data, ("common", "resources", "cpus"), (_is_number, cpus_msg))
+    errors = check(data, ("common", "resources", "cpus"), (is_number, cpus_msg))
     pids_msg = "common.resources.pids_limit must be an integer"
-    errors += check(data, ("common", "resources", "pids_limit"), (_is_int, pids_msg))
+    errors += check(data, ("common", "resources", "pids_limit"), (is_int, pids_msg))
     for field, example in _SIZE_FIELDS:
         message = f"common.resources.{field} must include a unit, e.g. '{example}'"
-        errors += check(data, ("common", "resources", field), (_is_size, message))
+        errors += check(data, ("common", "resources", field), (is_size, message))
     return errors
 
 
-def _is_string_list(value):
-    """True when *value* is a list whose entries are all strings."""
-    return isinstance(value, list) and all(isinstance(x, str) for x in value)
-
-
-def _check_starting(data):
+def _check_default(data):
     """users.starting, when present, must be a list of strings."""
-    return _optional(
+    return optional(
         data,
         ("users", "starting"),
-        (_is_string_list, "users.starting must be a list of strings"),
+        (is_string_list, "users.starting must be a list of strings"),
     )
 
 
 def _email_error(name, info):
     """users.<name>.email must be a valid address."""
-    if _is_email(info.get("email", "")):
+    if is_email(info.get("email", "")):
         return []
     return [f"users.{name}.email is not a valid email address"]
 
@@ -189,7 +88,7 @@ def _email_error(name, info):
 def _groups_error(name, info):
     """users.<name>.groups, when present, must be a list of strings."""
     groups = info.get("groups")
-    if groups is None or _is_string_list(groups):
+    if groups is None or is_string_list(groups):
         return []
     return [f"users.{name}.groups must be a list of strings"]
 
@@ -215,7 +114,7 @@ def _user_table_errors(name, info):
 
 def _check_user_tables(data):
     """Every [users.<name>] sub-table must hold a valid email and valid tags."""
-    users = _get(data, "users")
+    users = get_nested(data, "users")
     if not isinstance(users, dict):
         return []
     errors = []
@@ -226,15 +125,15 @@ def _check_user_tables(data):
 
 # Deployment-section fields checked when present: (section, key, predicate, label).
 _DEPLOY_FIELDS = (
-    ("frontend", "react-app-oauth-url", _is_url, "URL"),
-    ("localhost", "auth-authority", _is_url, "URL"),
-    ("localhost", "default-user", _is_username, "username"),
-    ("insecure-server", "oauth-url", _is_url, "URL"),
-    ("secure-server", "oauth-url", _is_url, "URL"),
-    ("workspace-localhost", "auth-authority", _is_url, "URL"),
-    ("workspace-localhost", "default-user", _is_username, "username"),
-    ("workspace-secure-server", "keycloak-issuer-url", _is_url, "URL"),
-    ("workspace-secure-server", "auth-authority", _is_url, "URL"),
+    ("frontend", "react-app-oauth-url", is_url, "URL"),
+    ("localhost", "auth-authority", is_url, "URL"),
+    ("localhost", "default-user", is_username, "username"),
+    ("insecure-server", "oauth-url", is_url, "URL"),
+    ("secure-server", "oauth-url", is_url, "URL"),
+    ("workspace-localhost", "auth-authority", is_url, "URL"),
+    ("workspace-localhost", "default-user", is_username, "username"),
+    ("workspace-secure-server", "keycloak-issuer-url", is_url, "URL"),
+    ("workspace-secure-server", "auth-authority", is_url, "URL"),
 )
 
 
@@ -243,7 +142,7 @@ def _check_deploy_fields(data):
     errors = []
     for section, key, predicate, label in _DEPLOY_FIELDS:
         message = f"{section}.{key} must be a valid {label}"
-        errors += _optional(data, (section, key), (predicate, message))
+        errors += optional(data, (section, key), (predicate, message))
     return errors
 
 
@@ -253,7 +152,7 @@ _CHECKS = (
     _check_path,
     _check_certs_src,
     _check_resources,
-    _check_starting,
+    _check_default,
     _check_user_tables,
     _check_deploy_fields,
 )
